requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.7
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.6
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
uritemplate==4.1.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
